feat: Enhance landing page with user details and date of birth input
- Added user name and date of birth fields to the form
- Integrated DatePicker for selecting date of birth
- Updated validation schema to include user name and date of birth
- Modified form submission logic to handle new fields
- Removed unused components and imports for cleaner code
- Improved local storage handling for user data
- Updated UI elements for better user experience
